+++ simple_manga_downloader/modules/__init__.py
@@ -3,4 +3,5 @@ from .mangaseeonline_us import Mangasee
 from .mangatown_com import Mangatown
 from .heavenmanga_org import Heavenmanga
 from .mangakakalot_com import Mangakakalot
+from .manganelo_com import Manganelo
 from .config_parser import Config